requirements-dev.txt
-r requirements.txt
pytest==9.0.2
pytest-flask==1.3.0
pytest-httpx==0.36.0
pytest-playwright==0.7.2
